added read and write file and more
added read and write file, added method for list fragment can not get
data to pass to it.  deleted Artist class file, placed into main to see
if that would fix issue, was able to pass data to the Artist class but
still can not figure out how to get data to pass through the list
fragment or detail fragment which was added. changed asyctask to return
jsonArray instead of object to see if that would work populate list as
well, will continue to work on to correct issues.